+++ app/scoreBundle.ts
@@ -5,6 +5,7 @@ import {recoverJSON} from "../inc/jsonRecovery";
 import {StaffToken, SheetDocument, PitchContext, PitchContextTable} from "../inc/staffSvg";
 import * as StaffNotation from "../inc/staffSvg/staffNotation";
 import * as SheetBaker from "./sheetBaker";
+import DictArray from "../inc/DictArray";
 
 
 
@@ -24,7 +25,7 @@ export default class ScoreBundle {
 
 
 	constructor (source, {onStatus = (..._) => _} = {}) {
-		this.scoreJSON = recoverJSON(source, {StaffToken, SheetDocument, PitchContext, PitchContextTable});
+		this.scoreJSON = recoverJSON(source, {StaffToken, SheetDocument, PitchContext, PitchContextTable, DictArray});
 		this.onStatus = onStatus;
 
 		if (!this.scoreJSON.midi)