+++ inc/staffSvg/index.ts
@@ -25,16 +25,12 @@ const parseSvgPage = (dom, ly, {logger = new LogRecorder(), attributes, ...optio
 };
 
 
-const recoverScoreJSON = text => recoverJSON(text, {StaffToken, SheetDocument});
-
-
 
 export {
 	svgToElements,
 	tokenizeElements,
 	organizeTokens,
 	parseSvgPage,
-	recoverScoreJSON,
 	StaffToken,
 	SheetDocument,
 	PitchContext,